+++ packages/myst-cli/src/process/mdast.ts
@@ -37,7 +37,7 @@ import {
   checkLinksTransform,
   embedTransform,
   importMdastFromJson,
-  includeFilesDirective,
+  includeFilesTransform,
   liftCodeMetadataToBlock,
   transformLinkedDOIs,
   transformOutputs,
@@ -52,6 +52,7 @@ import {
   reduceOutputs,
   transformPlaceholderImages,
   transformDeleteBase64UrlSource,
+  transformWebp,
 } from '../transforms/index.js';
 import type { ImageExtensions } from '../utils/index.js';
 import { logMessagesFromVFile } from '../utils/index.js';
@@ -99,6 +100,7 @@ export async function transformMdast(
     minifyMaxCharacters?: number;
     index?: string;
     simplifyFigures?: boolean;
+    processThumbnail?: boolean;
   },
 ) {
   const {
@@ -115,6 +117,7 @@ export async function transformMdast(
     minifyMaxCharacters,
     index,
     simplifyFigures,
+    processThumbnail,
   } = opts;
   const toc = tic();
   const { store, log } = session;
@@ -152,7 +155,7 @@ export async function transformMdast(
   cache.$internalReferences[file] = state;
   // Import additional content from mdast or other files
   importMdastFromJson(session, file, mdast);
-  includeFilesDirective(session, file, mdast);
+  includeFilesTransform(session, file, mdast, vfile);
   // This needs to come before basic transformations since it may add labels to blocks
   liftCodeMetadataToBlock(session, file, mdast);
 
@@ -216,15 +219,17 @@ export async function transformMdast(
       altOutputFolder: imageAltOutputFolder,
       imageExtensions,
     });
-    // Note, the thumbnail transform must be **after** images, as it may read the images
-    await transformThumbnail(session, mdast, file, frontmatter, imageWriteFolder, {
-      altOutputFolder: imageAltOutputFolder,
-      webp: !simplifyFigures,
-    });
-    await transformBanner(session, file, frontmatter, imageWriteFolder, {
-      altOutputFolder: imageAltOutputFolder,
-      webp: !simplifyFigures,
-    });
+    if (processThumbnail) {
+      // Note, the thumbnail transform must be **after** images, as it may read the images
+      await transformThumbnail(session, mdast, file, frontmatter, imageWriteFolder, {
+        altOutputFolder: imageAltOutputFolder,
+        webp: extraTransforms?.includes(transformWebp),
+      });
+      await transformBanner(session, file, frontmatter, imageWriteFolder, {
+        altOutputFolder: imageAltOutputFolder,
+        webp: extraTransforms?.includes(transformWebp),
+      });
+    }
   }
   await transformDeleteBase64UrlSource(mdast);
   const sha256 = selectors.selectFileInfo(store.getState(), file).sha256 as string;