+++ packages/myst-frontmatter/tests/examples.spec.ts
@@ -26,10 +26,11 @@ const files = [
   'affiliations.yml',
   'authors.yml',
   'credit.yml',
-  'orcid.yml',
-  'licenses.yml',
   'exports.yml',
+  'funding.yml',
   'keywords.yml',
+  'licenses.yml',
+  'orcid.yml',
   'thebe.yml',
 ];
 
@@ -76,6 +77,9 @@ casesList.forEach(({ title, frontmatter, cases }) => {
         }
         expect(opts.messages.warnings?.length ?? 0).toBe(warnings ?? 0);
         expect(opts.messages.errors?.length ?? 0).toBe(errors ?? 0);
+
+        // Test that normalized frontmatter is idempotent
+        expect(validator(normalized, opts)).toEqual(normalized);
       },
     );
   });