+++ packages/myst-to-tex/src/index.ts
@@ -15,19 +15,19 @@ import {
 } from './utils.js';
 import MATH_HANDLERS, { withRecursiveCommands } from './math.js';
 import { select, selectAll } from 'unist-util-select';
-import type { DefinitionDescription, DefinitionTerm, FootnoteDefinition } from 'myst-spec-ext';
+import type { FootnoteDefinition } from 'myst-spec-ext';
 import { transformLegends } from './legends.js';
 
 export type { LatexResult } from './types.js';
 
 const glossaryReferenceHandler: Handler = (node, state) => {
   if (!state.printGlossary) {
-    state.renderChildren(node);
+    state.renderChildren(node, true);
     return;
   }
 
   if (!node.identifier) {
-    state.renderChildren(node);
+    state.renderChildren(node, true);
     return;
   }
 
@@ -37,8 +37,11 @@ const glossaryReferenceHandler: Handler = (node, state) => {
       node,
       source: 'myst-to-tex',
     });
+    const gn = node as GenericNode;
+    state.write(toText(node).trim() || gn.label || '');
     return;
   }
+
   state.write('\\gls{');
   state.write(node.identifier);
   state.write('}');
@@ -52,9 +55,7 @@ const createFootnoteDefinitions = (tree: Root) =>
     }),
   );
 
-const createGlossaryDefinitions = (
-  tree: Root,
-): Record<string, [DefinitionTerm, DefinitionDescription]> =>
+const createGlossaryDefinitions = (tree: Root): Record<string, [string, string]> =>
   Object.fromEntries(
     selectAll('glossary > definitionList > *', tree)
       .map((node, i, siblings) => {
@@ -69,7 +70,9 @@ const createGlossaryDefinitions = (
         if (dd === undefined || dd.type !== 'definitionDescription') {
           throw new Error(`Definition term has no associated description`);
         }
-        return [dt.identifier, [dt, dd]];
+        const termText = toText(dt);
+        const descriptionText = toText(dd);
+        return [dt.identifier, [termText, descriptionText]];
       })
       .filter((x) => x.length > 0), // remove empty
   );
@@ -333,6 +336,10 @@ const handlers: Record<string, Handler> = {
       glossaryReferenceHandler(node, state, parent);
       return;
     }
+    // Note: if the md doc references a term not defined in the glossary,
+    //       the mdast will not have kind=definitionTerm and the logic
+    //       will follow this branch
+
     // Look up reference and add the text
     const usedTemplate = node.template?.includes('%s') ? node.template : undefined;
     const text = (usedTemplate ?? toText(node))?.replace(/\s/g, '~') || '%s';
@@ -408,7 +415,7 @@ class TexGlossaryAndAcronymSerializer {
   printedDefinitions: string;
 
   constructor(
-    glossaryDefinitions: Record<string, [DefinitionTerm, DefinitionDescription]>,
+    glossaryDefinitions: Record<string, [string, string]>,
     acronymDefinitions: Record<string, [string, string]>,
   ) {
     this.printedDefinitions = this.renderGlossary();
@@ -448,21 +455,26 @@ class TexGlossaryAndAcronymSerializer {
   }
 
   private createGlossaryDirectives(
-    glossaryDefinitions: Record<string, [DefinitionTerm, DefinitionDescription]>,
+    glossaryDefinitions: Record<string, [string, string]>,
   ): string[] {
     const directives = Object.keys(glossaryDefinitions).map((k) => ({
       key: k,
-      name: (glossaryDefinitions[k][0].children[0] as GenericNode).value || '',
-      description: (glossaryDefinitions[k][1].children[0] as GenericNode).value || '',
+      name: glossaryDefinitions[k][0],
+      description: glossaryDefinitions[k][1],
     }));
 
-    return directives.map(
+    const usepackage = '\\usepackage{glossaries}';
+    const makeglossaries = '\\makeglossaries';
+    const entries = directives.map(
       (entry) =>
         `\\newglossaryentry{${entry.key}}{name=${entry.name},description={${entry.description}}}`,
     );
+    return [usepackage, makeglossaries].concat(entries);
   }
 
-  private createAcronymDirectives(acronymDefinitions: Record<string, [string, string]>): string[] {
+  private createAcronymDirectives(
+    acronymDefinitions: Record<string, [string, string]>
+  ): string[] {
     const directives = Object.keys(acronymDefinitions).map((k) => ({
       key: k,
       acronym: acronymDefinitions[k][0],
@@ -482,7 +494,7 @@ class TexSerializer implements ITexSerializer {
   handlers: Record<string, Handler>;
   references: References;
   footnotes: Record<string, FootnoteDefinition>;
-  glossary: Record<string, [DefinitionTerm, DefinitionDescription]>;
+  glossary: Record<string, [string, string]>;
   abbreviations: Record<string, [string, string]>;
 
   constructor(file: VFile, tree: Root, opts?: Options) {