+++ packages/myst-common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # myst-common
 
+## 1.1.6
+
 ## 1.1.5
 
 ### Patch Changes