+++ packages/myst-config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # myst-config
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [911d1b1]
+- Updated dependencies [911d1b1]
+- Updated dependencies [911d1b1]
+- Updated dependencies [59b54584]
+  - myst-frontmatter@1.1.6
+
 ## 1.1.5
 
 ### Patch Changes